refactored the server code so that it's easier to understand
