+++ package.json
@@ -114,6 +114,7 @@
     "karma-webpack": "1.8.1",
     "mocha": "2.3.4",
     "moment": "2.12.0",
+    "on-build-webpack": "^0.1.0",
     "selenium-webdriver": "3.0.0-beta-2",
     "should": "8.0.2"
   },
