+++ server/scripts/exchangerates.js
